`t.fail(err.message)`
When calling `t.fail` with the whole `Error` object, all keys are printed. When the `err` has a `stream` property or similar this can be massively noisy to the point of overflowing GHA's default max log length for viewing in the browser.

While some error properties are useful (`stack`, `code`) this is the most straightforward way to cut the noise.